🚚 move benchmarking code to test file
