+++ src/CMakeLists.txt
@@ -11,7 +11,6 @@ add_library(
   checker/dd/DDConstructionChecker.cpp
   checker/dd/DDSimulationChecker.cpp
   checker/dd/DDAlternatingChecker.cpp
-  checker/dd/DDPartialEquivalenceBenchmarks.cpp
   checker/dd/applicationscheme/GateCostApplicationScheme.cpp
   checker/dd/simulation/StateGenerator.cpp
   checker/zx/ZXChecker.cpp)
